requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Document processing
pypdf>=3.12.0
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
chardet>=5.0.0

# ML and NLP
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Vector search and retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# API and web services
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0

# LLM integrations
openai>=1.0.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Data serialization
jsonlines>=3.1.0

# Optional: Jupyter support
jupyter>=1.0.0
ipykernel>=6.25.0
